+++ R/data.R
@@ -1053,9 +1053,9 @@
 #' Servers maintained by the Baltimore City Mayor's Office of Information
 #' Technology (MOIT) Enterprise GIS (EGIS) program. A limited number of
 #' potential sensitive and unresponsive server layers have been excluded. Used
-#' by the [get_baltimore_esri_data()] function. Updated December 23, 2022.
+#' by the [get_baltimore_esri_data()] function. Updated 2023 May 26.
 #'
-#' @format A data frame with 1,286 rows and 17 variables:
+#' @format A data frame with 1,324 rows and 17 variables:
 #' \describe{
 #'   \item{`name`}{Name}
 #'   \item{`nm`}{Name with snake case}